Move model from experimental to demos
