+++ models/demos/yolov8s/README.md
@@ -14,27 +14,27 @@ YOLOv8 is one of the recent iterations in the YOLO series of real-time object de
 
 ## Details
 The entry point to yolov8s model is YOLOv8s in
-`models/experimental/yolov8s/tt/ttnn_yolov8s.py`.
+`models/demos/yolov8s/tt/ttnn_yolov8s.py`.
 
 ### Use the following commands for 640x640 to run the :
 
 - yolov8s demo :
 ```bash
-pytest models/experimental/yolov8s/demo/demo.py
+pytest models/demos/yolov8s/demo/demo.py
 ```
 
-- e2e perf with trace(100 fps):
+- e2e perf with trace(165 fps):
 ```bash
-pytest models/experimental/yolov8s/tests/test_e2e_performant.py
+pytest models/demos/yolov8s/tests/test_e2e_performant.py
 ```
 
 #### Note: The post-processing is performed using PyTorch.
 
 ## Inputs
-The demo receives inputs from `models/experimental/yolov8s/demo/images` dir by default. To test the model on different input data, it is recommended to add a new image file to this directory.
+The demo receives inputs from `models/demos/yolov8s/demo/images` dir by default. To test the model on different input data, it is recommended to add a new image file to this directory.
 
 ## Outputs
-A runs folder will be created inside the `models/experimenatl/yolov8s/demo/` directory. For reference, the model output will be stored in the torch_model directory, while the TTNN model output will be stored in the tt_model directory.
+A runs folder will be created inside the `models/demos/yolov8s/demo/` directory. For reference, the model output will be stored in the torch_model directory, while the TTNN model output will be stored in the tt_model directory.
 
 ## Additional Information:
 The tests can be run with  randomly initialized weights and pre-trained real weights.  To run only for the pre-trained weights, specify pretrained_weight_true when running the tests.
