+++ models/demos/yolov8s/demo/images/bus.jpg
