+++ models/demos/yolov8s/tt/tt_yolov8s_utils.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import torch
+
 import ttnn
