+++ tests/scripts/single_card/run_single_card_demo_tests.sh
@@ -237,7 +237,7 @@ run_yolov8x_perf() {
 }
 run_yolov8s_perf() {
 
-  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto --disable-warnings models/experimental/yolov8s/demo/demo.py --timeout 600
+  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto --disable-warnings models/demos/yolov8s/demo/demo.py --timeout 600
 
 }
